Create GUI.py
This is the GUI for the program. It creates an instance of the board class then draws the board and pieces on screen accordingly.

The user cannot interact with the GUI yet but it does have a starting animation. If you are testing the program and wish to load it faster, disable the starting animation by changing boolean variable animation to False in the main.