Created game.py and player.py
Created Game class with an init and several methods that will be
required to play the game, and included a main block that starts the
game and initializes the gui. Imported all necessary classes into game.
Also created the player class which will hold all the information
that defines a player in our chess game.
Changed the name of a variable in board.py to be consistent with our
naming convention.
A few of the methods in the game class are incomplete and TODO comments
have been placed appropriately.